# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Reportes",
        "contact": {
            "email": "antony.rodriguez@mimco.com.pe"
        },
        "version": "1"
    },
    "paths": {
        "/GestionReportes/public/index.php/segu_grup": {
            "post": {
                "tags": [
                    "Seguimiento Grupos"
                ],
                "summary": "Permite listar los grupos para visualizar su avance.",
                "operationId": "App\\Http\\Controllers\\GruposController::segu_grup",
                "parameters": [
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "Codigo de Proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 175
                    },
                    {
                        "name": "intIdProyTarea",
                        "in": "path",
                        "description": "Codigo Programa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": -1
                    },
                    {
                        "name": "intIdProyPaquete",
                        "in": "path",
                        "description": "Codigo Grupo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": -1
                    },
                    {
                        "name": "intIdArmadores",
                        "in": "path",
                        "description": "Codigo Armadores",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": -1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "varNumeDni": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProy": "175",
                                    "intIdTipoProducto": "1",
                                    "intIdProyTarea": "-1",
                                    "intIdProyPaquete": "-1",
                                    "intIdArmadores": "-1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sin Mensaje"
                    },
                    "407": {
                        "description": "El Documento de identidad ingresado no se encuentra registrado."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
    }
}